try new way for string concatination and docker container name generation

+++ app/rancher.sh
@@ -1,15 +1,21 @@
 
 #load container name from rancher meta-data
-CONTAINER_NAME="r-$(curl -s rancher-metadata/latest/self/container/name)"
-NGINX_DOCKER_GEN_CONTAINER=${CONTAINER_NAME/letsencrypt/nginx-gen}
-NGINX_DOCKER_CONTAINER=${CONTAINER_NAME/letsencrypt_/}
+STACK_NAME="r-$(curl -s rancher-metadata/latest/self/stack/name)"
+SCALE="$(curl -s rancher-metadata/latest/self/service/scale)"
+
+NGINX_DOCKER_GEN_CONTAINER=$STACK_NAME"_nginx_nginx-gen_"$SCALE
+NGINX_DOCKER_CONTAINER=$STACK_NAME"_nginx_"$SCALE
 
 export NGINX_DOCKER_GEN_CONTAINER
 export NGINX_DOCKER_CONTAINER
 
 #download docker-gen nginx-template
-curl https://raw.githubusercontent.com/jwilder/nginx-proxy/master/nginx.tmpl > /etc/nginx/docker-gen/nginx.tmpl
+if [ ! -f /etc/nginx/docker-gen/nginx.tmpl ]; then
+    curl https://raw.githubusercontent.com/jwilder/nginx-proxy/master/nginx.tmpl > /etc/nginx/docker-gen/nginx.tmpl
+fi
 
 #create docker-gen config file
-echo "$NGINX_DOCKER_CONTAINER = 1" >> /app/docker-gen.cfg
-cp /app/docker-gen.cfg /etc/nginx/docker-gen/docker-gen.cfg
+if [ ! -f /etc/nginx/docker-gen/docker-gen.cfg ]; then
+    echo "$NGINX_DOCKER_CONTAINER = 1" >> /app/docker-gen.cfg
+    cp /app/docker-gen.cfg /etc/nginx/docker-gen/docker-gen.cfg
+fi
